+++ packages/nuxt/src/app/composables/payload.ts
@@ -1,6 +1,6 @@
 import { joinURL } from 'ufo'
 import { useNuxtApp } from '../nuxt'
-import { useRuntimeConfig } from '..'
+import { useRuntimeConfig } from '../nuxt'
 import { useHead } from './head'
 
 interface LoadPayloadOptions {
